feat: replace ibisibi.service with install script

+++ deploy.sh
@@ -14,15 +14,18 @@ CROSS_TRIPLE="arm-unknown-linux-gnueabihf"
 cd $(cd -P -- "$(dirname -- "$0")" && pwd -P) && \
 
 # And build for CROSS_TRIPLE
-./build.sh $CROSS_TRIPLE || exit 1
+RELEASE_DIR_NAME=$(./build.sh $CROSS_TRIPLE)
+if test -z $RELEASE_DIR; then
+  echo -e "Build failed, exiting..."
+  exit 1
+fi
 
 # Copy to Raspberry Pi via scp
-echo "Build successful, deploying binary to $RPI_HOSTNAME..."
+echo "Build successful, deploying release to $RPI_HOSTNAME..."
 
 # Stop service for upgrade first (if running)
 ssh $SSH_HOST 'sudo systemctl stop ibisibi'
 
 # Update unit file and executable, restart service
-scp ibisibi.service $SCP_DEST && \
-scp target/$CROSS_TRIPLE/release/ibisibi $SCP_DEST && \
-ssh $SSH_HOST 'rm ~/bin/ibisibi; ln -s ~/ibisibi/ibisibi ~/bin/ibisibi && sudo mv -f ~/ibisibi/ibisibi.service /etc/systemd/system && sudo systemctl enable ibisibi && sudo systemctl start ibisibi'
+scp -r release/$RELEASE_DIR_NAME $SCP_DEST && \
+ssh $SSH_HOST 'cd $RELEASE_DIR_NAME && ./install.sh'
